add sequence.go
